LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-in-shadow-002-expected.txt


FAIL Anchor names in different tree scopes should not be confused assert_equals: expected 100 but got 0
